Moved @tryghost/mentions-email-report to Ghost core
fix https://linear.app/ghost/issue/ENG-2069/mentions-email-report

- this commit migrates the code and test for the mentions email report
  service into Ghost core so we can keep all related code together and
  reduce overhead
- as a result of this change, we can remove the package and associated
  references
