+++ compose.yml
@@ -79,7 +79,6 @@ services:
       - node_modules_ghost_members-csv:/home/ghost/ghost/members-csv/node_modules:delegated
       - node_modules_ghost_members-importer:/home/ghost/ghost/members-importer/node_modules:delegated
       - node_modules_ghost_members-ssr:/home/ghost/ghost/members-ssr/node_modules:delegated
-      - node_modules_ghost_mentions-email-report:/home/ghost/ghost/mentions-email-report/node_modules:delegated
       - node_modules_ghost_metrics-server:/home/ghost/ghost/metrics-server/node_modules:delegated
       - node_modules_ghost_milestones:/home/ghost/ghost/milestones/node_modules:delegated
       - node_modules_ghost_minifier:/home/ghost/ghost/minifier/node_modules:delegated
@@ -251,7 +250,6 @@ volumes:
   node_modules_ghost_members-csv: {}
   node_modules_ghost_members-importer: {}
   node_modules_ghost_members-ssr: {}
-  node_modules_ghost_mentions-email-report: {}
   node_modules_ghost_metrics-server: {}
   node_modules_ghost_milestones: {}
   node_modules_ghost_minifier: {}
