+++ ghost/core/test/unit/server/services/mentions-email-report/mention-email-report-job.test.js
@@ -1,5 +1,5 @@
 const sinon = require('sinon');
-const MentionEmailReportJob = require('../');
+const MentionEmailReportJob = require('../../../../../core/server/services/mentions-email-report/MentionEmailReportJob');
 
 class MockMentionReportRecipientRepository {
     #recipients = [{
