--- ghost/mentions-email-report/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/MentionEmailReportJob');
